textfsm_test: do not relay on dic order (#33)
The test `testNestedMatching` compares a list of dictionaries with
a string.  In Python 2 (and some versions of Python 3) the order
of a dictionary is not guaranteed, and this makes the test to fail.

This patch compares the raw result with the raw list, and assetEqual
takes care of the ordering.